+++ webstack/libs/applications/src/lib/apps/Calculator/styling.css
