+++ webstack/libs/applications/src/lib/initialValues.ts
@@ -35,7 +35,7 @@ import { name as VegaLiteViewerName, init as defaultVegaLiteViewer } from './app
 import { name as VideoViewerName, init as defaultVideoViewer } from './apps/VideoViewer';
 import { name as WebpageLinkName, init as defaultWebpageLink } from './apps/WebpageLink';
 import { name as WebviewName, init as defaultWebview } from './apps/Webview';
-import { name as counterTestName, init as defaultcounterTest } from './apps/counterTest';
+import { name as CalculatorName, init as defaultCalculator } from './apps/Calculator';
 
 export const initialValues = {
   [AIPaneName]: defaultAIPane,
@@ -73,6 +73,6 @@ export const initialValues = {
   [VideoViewerName]: defaultVideoViewer,
   [WebpageLinkName]: defaultWebpageLink,
   [WebviewName]: defaultWebview,
-  [counterTestName]: defaultcounterTest,
+  [CalculatorName]: defaultCalculator,
 };
